Remove .expo extensions
