+++ packages/expo/src/launch/registerRootComponent.tsx
@@ -1,7 +1,31 @@
 import 'expo/build/Expo.fx';
+
+import Constants, { ExecutionEnvironment } from 'expo-constants';
 import * as React from 'react';
 import { AppRegistry } from 'react-native';
 
+if (__DEV__) {
+  // In dev mode, attempt to keep the screen on.
+  try {
+    const { activateKeepAwake } = require('expo-keep-awake') 
+    activateKeepAwake();
+  } catch {
+    // expo-keep-awake may not be installed in all projects.
+  }
+}
+
+const isExpo = Constants.executionEnvironment === ExecutionEnvironment.Standalone || Constants.executionEnvironment === ExecutionEnvironment.StoreClient;
+
 export default function registerRootComponent<P>(component: React.ComponentType<P>): void {
-  AppRegistry.registerComponent('main', () => component);
+
+  if (isExpo) {
+    try {
+      const withExpoRoot = require('./withExpoRoot');
+      AppRegistry.registerComponent('main', () => withExpoRoot(component));
+    } catch (error) {
+      if (isExpo) throw error;
+    }
+  } else {
+    AppRegistry.registerComponent('main', () => component);
+  }
 }
